refactor: Import service names from the module that owns them
services/__init__.py re-exported thirty names from six modules for five
consumers. It existed so the helper.py split could land without touching a
caller; that migration is over, and what is left is a second place every name
could have come from.
