+++ cogs/birthday.py
@@ -10,7 +10,7 @@ from discord.utils import escape_markdown
 from constants import MAX_DAYS, Months
 from db import repository
 from errors import notify, report
-from services import is_leap_day, next_birthday_on
+from services.birthday import is_leap_day, next_birthday_on
 from services.config import config, has_configured_role
 
 logger = logging.getLogger(__name__)
