+++ cogs/tasks.py
@@ -10,8 +10,9 @@ from discord.utils import escape_markdown
 from db import repository
 from db.models import DiscordUser
 from errors import notify, report
-from services import next_birthday, send_message
+from services.birthday import next_birthday
 from services.config import config
+from services.send import send_message
 from services.twitch.api import broken_subscriptions
 from services.twitch.helix import HelixError
