+++ services/__init__.py
@@ -1,59 +1,5 @@
-from . import audit
-from .birthday import is_leap_day, next_birthday, next_birthday_on
-from .duration import format_unit, get_age, get_ordinal_suffix
-from .present import get_channel_mention, get_discriminator, get_pfp
-from .roles import get_member_role, roles_button_pressed, toggle_role
-from .send import edit_embed, send_embed, send_message
-from .twitch.api import (
-    get_ad_schedule,
-    get_channel,
-    get_stream,
-    get_stream_vod,
-    get_subscriptions,
-    get_user,
-    get_user_by_username,
-    get_users,
-    subscribe_to_user,
-    unsubscribe_to_user,
-)
-from .twitch.commands import dispatch
-from .twitch.signature import (
-    get_hmac,
-    get_hmac_message,
-    parse_rfc3339,
-    verify_message,
-)
+"""What the bot does, as opposed to how it is reached.
 
-__all__ = [
-    "audit",
-    "dispatch",
-    "edit_embed",
-    "format_unit",
-    "get_ad_schedule",
-    "get_age",
-    "get_channel",
-    "get_channel_mention",
-    "get_discriminator",
-    "get_hmac",
-    "get_hmac_message",
-    "get_member_role",
-    "get_ordinal_suffix",
-    "get_pfp",
-    "get_stream",
-    "get_stream_vod",
-    "get_subscriptions",
-    "get_user",
-    "get_user_by_username",
-    "get_users",
-    "is_leap_day",
-    "next_birthday",
-    "next_birthday_on",
-    "parse_rfc3339",
-    "roles_button_pressed",
-    "send_embed",
-    "send_message",
-    "subscribe_to_user",
-    "toggle_role",
-    "unsubscribe_to_user",
-    "verify_message",
-]
+Nothing is re-exported here: import from the module that owns the name, so a
+reader lands on the file that defines it rather than on a list of thirty.
+"""
